g28-extraction-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdf2image==1.17.0
Pillow==10.2.0
anthropic==0.40.0
pydantic==2.6.0
